requirements.txt
streamlit
tensorflow>=2.17
Pillow
numpy
requests
